Implement "Internet Zone" support
This adds a new configuration section "ZonePatterns" that enable us to
set the default browser for each internet zone.

    [ZonePatterns]
    0001=intra|ie
    0002=internet|firefox

Note that the current implementation checks patterns in the following
order: URLPatterns > HostNamePatterns > ZonePatterns.

This means that we can make some exceptions to zone rules using URL (or
HostName) patterns, which enables users to say "We want to use Firefox
for internet, but meanwhile we want to use IE for http://example.com/".

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>